+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- fixed a possible TypeError with the `ACP3\Core\Picture\Output` class
+
 ## [4.30.2] - 2018-10-04
 ### Fixed
 - fixed a possible TypeError when creating a download